Extract captcha logic to abstract base class
