+++ angular/src/components/login.component.ts
@@ -19,8 +19,9 @@ import { StorageService } from 'jslib-common/abstractions/storage.service';
 
 import { ConstantsService } from 'jslib-common/services/constants.service';
 
-import { CaptchaIFrame } from 'jslib-common/misc/captcha_iframe';
 import { Utils } from 'jslib-common/misc/utils';
+
+import { CaptchaProtectedComponent } from './captchaProtectedComponent';
 
 const Keys = {
     rememberedEmail: 'rememberedEmail',
@@ -28,15 +29,12 @@ const Keys = {
 };
 
 @Directive()
-export class LoginComponent implements OnInit {
+export class LoginComponent extends CaptchaProtectedComponent implements OnInit {
     @Input() email: string = '';
     @Input() rememberEmail = true;
 
     masterPassword: string = '';
     showPassword: boolean = false;
-    captchaSiteKey: string = null;
-    captchaToken: string = null;
-    captcha: CaptchaIFrame;
     formPromise: Promise<AuthResult>;
     onSuccessfulLogin: () => Promise<any>;
     onSuccessfulLoginNavigate: () => Promise<any>;
@@ -46,10 +44,12 @@ export class LoginComponent implements OnInit {
     protected successRoute = 'vault';
 
     constructor(protected authService: AuthService, protected router: Router,
-        protected platformUtilsService: PlatformUtilsService, protected i18nService: I18nService,
-        protected stateService: StateService, protected environmentService: EnvironmentService,
+        platformUtilsService: PlatformUtilsService, i18nService: I18nService,
+        protected stateService: StateService, environmentService: EnvironmentService,
         protected passwordGenerationService: PasswordGenerationService,
-        protected cryptoFunctionService: CryptoFunctionService, private storageService: StorageService) { }
+        protected cryptoFunctionService: CryptoFunctionService, private storageService: StorageService) {
+        super(environmentService, i18nService, platformUtilsService);
+    }
 
     async ngOnInit() {
         if (this.email == null || this.email === '') {
@@ -66,19 +66,7 @@ export class LoginComponent implements OnInit {
             this.focusInput();
         }
 
-        let webVaultUrl = this.environmentService.getWebVaultUrl();
-        if (webVaultUrl == null) {
-            webVaultUrl = 'https://vault.bitwarden.com';
-        }
-        this.captcha = new CaptchaIFrame(window, webVaultUrl,
-            this.i18nService, (token: string) => {
-                this.captchaToken = token;
-            }, (error: string) => {
-                this.platformUtilsService.showToast('error', this.i18nService.t('errorOccurred'), error);
-            }, (info: string) => {
-                this.platformUtilsService.showToast('info', this.i18nService.t('info'), info);
-            }
-        );
+        this.setupCaptcha();
     }
 
     async submit() {
@@ -107,9 +95,8 @@ export class LoginComponent implements OnInit {
             } else {
                 await this.storageService.remove(Keys.rememberedEmail);
             }
-            if (!Utils.isNullOrWhitespace(response.captchaSiteKey)) {
-                this.captchaSiteKey = response.captchaSiteKey;
-                this.captcha.init(response.captchaSiteKey);
+            if (this.handleCaptchaRequired(response)) {
+                return;
             } else if (response.twoFactor) {
                 if (this.onSuccessfulLoginTwoFactorNavigate != null) {
                     this.onSuccessfulLoginTwoFactorNavigate();
@@ -165,9 +152,6 @@ export class LoginComponent implements OnInit {
             '&state=' + state + '&codeChallenge=' + codeChallenge);
     }
 
-    showCaptcha() {
-        return !Utils.isNullOrWhitespace(this.captchaSiteKey);
-    }
     protected focusInput() {
         document.getElementById(this.email == null || this.email === '' ? 'email' : 'masterPassword').focus();
     }
